refactor: 디자인 변경에 따른 기존 no-sidebar 내부 페이지 사이드바 있는 폴더 내부로 이동 (#45)
