+++ src/app/admin/(with-sidebar)/applications/[id]/_components/AdminReviewTabs.tsx
