+++ src/app/admin/(with-sidebar)/applications/[id]/_components/BasicInfoView.tsx
