+++ src/constants/routes.ts
@@ -11,7 +11,7 @@ export const ROUTES = {
 
   //admin routes
   ADMIN: '/admin',
-  ADMIN_APPLICATION: '/admin/applications',
+  ADMIN_APPLICATIONS: '/admin/applications',
   ADMIN_APPLICATION_EDIT: '/admin/application-edit',
   ADMIN_RECRUITMENT: '/admin/recruitment',
   ADMIN_RESULTS: '/admin/results',
